Fix GetJuliaIndent() messing with "ignorecase" option
Issue #13 was about ignorecase messing with the Julia indentation, and
the fix was to disable ignorecase during the execution of GetJuliaIndent().

In one instance it wasn't enabled at the end of GetJuliaIndent(). If the
following file is edited:

    f(
    )

and == is hit on the second line, then the 'ignorecase' option will
remain disabled.

After this fix, it will be properly enabled.