Special parse of import/using/export; as keyword
Should be robust enough, and not too bad performance-wise,
except for lists with hundreds of names, one per line,
like e.g. base/exports.jl

The "as" keyword is also recognized in export lines even though it
shouldn't. This is intended to unify indentation of
import/using/export lines and should be mostly harmless.