Improve inner-block selection
Addresses part of #286, dealing with multi-line brackets.

Still doesn't work well when the block header is multi-line because of
other types of incomplete expressions, e.g. because of dangling binary
operators at the end of a line:

  if x + y ==
      z
      println("yes")
  end

Another case which is not covered is if code continues in the same line
as the header:

  if x println("yes")
      return
  end